Added entry point and notes on uploading to PyPI.
