+++ .gitignore
@@ -12,3 +12,5 @@ howto_gmsh/*.msh
 howto_mesh_terrain/*.vtu
 howto_mesh_terrain/*.msh
 howto_mesh_terrain/*.stl
+__pycache__
+.venv
