+++ setup.cfg
@@ -20,3 +20,6 @@ packages = find:
 python_requires = >=3.8
 install_requires = meshio 
 
+[options.entry_points]
+console_scripts =
+    msh2vtu = msh2vtu:main
